React router properly configured and PDP added
