+++ src/main.tsx
@@ -4,11 +4,14 @@ import App from "./App.tsx";
 import "./index.css";
 import { ApolloProvider } from "@apollo/client";
 import client from "./lib/apollo/client.ts";
+import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ApolloProvider client={client}>
-      <App />
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
     </ApolloProvider>
   </React.StrictMode>,
 );
