+++ src/pages/Main/Main.module.scss
