+++ src/pages/ProductDetails/ProductDetails.tsx
@@ -1,0 +1,9 @@
+const ProductDetails = () => {
+  return (
+    <div>
+      <h1>PRODUCT DETAILS PAGE</h1>
+    </div>
+  );
+};
+
+export default ProductDetails;
